Generates a Design object with the given specifications.

Generates a randomized control treatment Design (rct_design()), or an observational Design (obs_design()), or a regression discontinuity Design (rd_design()).

Usage

rct_design(formula, data, subset = NULL, dichotomy = NULL, na.fail = TRUE)

rd_design(formula, data, subset = NULL, dichotomy = NULL, na.fail = TRUE)

obs_design(formula, data, subset = NULL, dichotomy = NULL, na.fail = TRUE)

Arguments

formula

defines the Design components

data

the data set.

subset

optional, subset the data before creating the Design object

dichotomy

optional, a formula defining the dichotomy of the treatment variable if it isn't already 0/1 or logical. See the dichotomy() function for details.

na.fail

If TRUE (default), any missing data found in the variables specified in formula (excluding treatment) will trigger an error. If FALSE, non-complete cases will be dropped before the creation of the Design

Value

a Design object of the requested type for use in further analysis

Details

The formula must include exactly one of unit_of_assignment(), uoa(), cluster(), or unitid() to identify the units of assignment (one or more variables). These are all synonyms; the choice of which has no impact on the analysis. If defining an rd_design, the formula must also include a forcing() entry. The formula may optionally include a block() as well. Each of these can take in multiple variables, e.g. to pass both a hosuehold ID and individual ID as unit of assignment, use uoa(hhid, iid) and not uoa(hhid) + uoa(iid).

The treatment variable passed into the left-hand side of formula can either be logical, numeric, or character. If it is anything else, it is attemped to be converted to one of those (for example, factor and ordered are converted to numeric if the levels are numeric, otherwise to character. If the treatment is not logical or numeric with only values 0 and 1, in order to generate weights with ate() or ett(), the dichotomy argument must be used to identify the treatment and control groups. The Design creation functions (rct_design(), rd_design(), obs_design()) all support the dichotomy argument, or instead dichotomy can be passed to ett() and ate() directly. See dichotomy() for more details on specifying a dichotomy.

Examples

data(simdata)
des <- rct_design(z ~ unit_of_assignment(cid1, cid2) + block(bid), data = simdata)

data(schooldata)
des <- obs_design(treatment ~ unit_of_assignment(schoolid) + block(state),
                  data = schooldata)