Concatenate weights

Given several variations of weights generated from a single Design, combine into a single weight.

Usage

# S4 method for WeightedDesign
c(x, ..., force_dichotomy_equal = FALSE)

Arguments

x,

.. a WeightedDesign object, typically created from ate() or ett()

...

any number of additional WeightedDesign objects with equivalent Design to x and eachother

force_dichotomy_equal

if FALSE (default), Designs are considered equivalent even if their dichotomy differs. If TRUE, @dichotomy must also be equal.

Value

If all WeightedDesign objects contain the same dichotomy (including NULL), a new WeightedDesign with the weights concatenated in the input order.

If force_dichotomy_equal is FALSE and the WeightedDesign objects differ in their dichotomies, a CombinedWeightedDesign is returned. This functions essentially identically to a WeightedDesign, but carries with it the dichotimization information that came from each of the individual WeightedDesign.

Details

Concatenating WeightedDesign objects with c() requires both individual WeightedDesign objects to come from the same Design (except dichotomy, see below) and have the same target (e.g all created with ate() or all created with ett(), no mixing-and-matching). All arguments to c() must be WeightedDesign.

One exception is when concatenting WeightedDesign objects whose Design differ only in their dichotomies. There may be cases where the treatment is continuous or has multiple levels, and there is a need to combine the weights from the same general design, but with different dichotomies. Therefore multiple WeightedDesign objects can be combined if they are identical except for their @dichotomy slots. The resulting object will be a CombinedWeightedDesign which tracks all individual dichotomy.

Examples

data(simdata)
des <- rct_design(z ~ unit_of_assignment(uoa1, uoa2), data = simdata)
w1 <- ate(des, data = simdata[1:30,])
w2 <- ate(des, data = simdata[31:40,])
w3 <- ate(des, data = simdata[41:50,])
c_w <- c(w1, w2, w3)
c(length(w1), length(w2), length(w3), length(c_w))
#> [1] 30 10 10 50

des <- rct_design(dose ~ unit_of_assignment(uoa1, uoa2), data = simdata)
w1 <- ate(des, data = simdata[1:10, ], dichotomy = dose >= 300 ~ .)
w2 <- ate(des, data = simdata[11:30, ], dichotomy = dose >= 200 ~ .)
w3 <- ate(des, data = simdata[31:50, ], dichotomy = dose >= 100 ~ .)
c_w <- c(w1, w2, w3)