Extract empirical estimating equations from a DirectAdjusted model fit

Extract empirical estimating equations from a DirectAdjusted model fit

Extract bread matrix from a DirectAdjusted model fit

Usage

# S3 method for DirectAdjusted
estfun(x, ...)

# S3 method for DirectAdjusted
bread(x, ...)

Arguments

x

a fitted DirectAdjusted object

...

arguments passed to methods

Value

An \(n\times k\) matrix of empirical estimating equations for the covariance-adjusted ITT effect regression. \(n\) represents the number of observations in the union of the samples used to fit the two regressions. \(k\) represents the number of parameters in the latter model.

Each row represents an observation's contribution to the stacked estimating equations. This contribution, denoted \(\tilde{\psi}_{i}\) for the \(i\)th observation, is given by $$\tilde{\psi}_{i} = \psi_{i} + \phi_{i}A_{11}^{-1}A_{21}^{T}$$ where \(\psi_{i}\) is the observation's contribution to the ITT effect model fit, \(\phi_{i}\) is the observation's contribution to the covariance adjustment model fit, and the \(A\) matrices are given by typical sandwich calculations.

Note that the formulation of the output matrix \(\tilde{\Psi}\) requires information about each observation's contributions to both the covariance adjustment and ITT effect models (where some observations may not contribute to both models). Estimating equations are taken from sandwich::estfun

calls on both fitted models and aligned as closely as possible. The by argument in cov_adj() can be used to specify a column unrelated to the design that will allow for exact alignment of such matrices. If no by

argument is provided, clustering information given in the DirectAdjusted's Design object will be used to align rows by unit of assignment, even though no guarantees can be made about aligning the matrices within units of assignment. Regardless of the eventual alignment and initial ordering of the observations in the two matrices, however, when using vcovDA, variance estimates will ultimately be the same due to the clustering passed to any sandwich::meatCL calls.

Details

This function is a thin wrapper around .get_tilde_a22_inverse().