(Internal) Locate a Design in the call stack

assigned()/ate()/ett()/cov_adj() all need the Design to operate. If any are called in the model without a design= argument, this function sees if it can find the Design in another of these functions.

Usage

.get_design(NULL_on_error = FALSE)

Arguments

NULL_on_error

if TRUE, returns NULL if a Design object is not found rather than throwing an error.

Value

A Design, or NULL if NULL_on_error is TRUE and the Design can't be found.

Details

Note that it will never look inside assigned() (gets complicated in formulas), only in weights or cov_adj(). E.g.

lm(y ~ assigned(), weights = ate(des), offest = cov_adj(mod1))

lm(y ~ assigned(), weights = ate(), offest = cov_adj(mod1, design = des))

will both work, but

lm(y ~ assigned(des), weights = ate(), offest = cov_adj(mod1))

will fail.