Covariance Adjustment for Treatment Estimation

Prior to obtaining predicted values, cov_adj() tries to identify the treatment variable (as specified in the design) and replace it with the reference level. If the treatment is binary, this is FALSE. If treatment is numeric, it is the smallest non-negative value (note that this means for 0/1 binary, it uses a 0). Factor treatments are not currently supported, but if we add them, it will use the first level() of the factor, you may change this by using relevel() to adjust.

Usage

cov_adj(model, newdata = NULL, design = NULL, by = NULL)

Arguments

model

Any model that inherits from a glm, lm, or robustbase::lmrob object

newdata

Optional; a data.frame of new data

design

Optional Design. If not provided, the function will search through the call stack to find one.

by

optional; vector or list connecting names of unit of assignment/unitid/cluster variables in design to unit of assignment/unitid/cluster variables in the covariance adjustment data. If a named vector, names should represent variables in the Design object and values should represent variables in the data. Only needed if:

columns not related to the design should be used for merging the covariance adjustment and quasiexperimental samples, or 2) the column names differ between the datasets used to fit both models.

Value

SandwichLayer or PreSandwichLayer object; the former if design is provided or a design can be found in the call stack, otherwise the latter. The values represent the covariance adjustments for the observations in newdata, if newdata is provided or found as an argument to lmitt.formula, or the fitted values from model. The length of the output of cov_adj() varies depending on this logic.

Examples

data(STARdata)
STARdata$treatment <- STARdata$stark == "small"
STARdata$treatment[is.na(STARdata$treatment)] <- FALSE
STARdata$studentid <- as.character(seq_len(nrow(STARdata)))

covariance_y0_read <- lm(readk ~ gender + ethnicity + lunchk +
                                 ladderk + experiencek + tethnicityk,
                         data = STARdata, subset = !treatment)

STARdata_design <- rct_design(treatment ~ cluster(studentid), data = STARdata)
STARdata_ate    <- ate(STARdata_design, data = STARdata)
STARdata_ett    <- ett(STARdata_design, data = STARdata)

ett_read <- lm(readk ~ treatment,
               offset = cov_adj(covariance_y0_read, newdata = STARdata),
               data = STARdata,
               weights = STARdata_ett)
#> Warning: Some covariance adjustments are NA; be careful of dropping these observations when fitting the ITT effect model
coef(ett_read)
#>   (Intercept) treatmentTRUE 
#> -8.224997e-13  5.554179e+00 

ate_read <- lm(readk ~ treatment,
               offset = cov_adj(covariance_y0_read, newdata = STARdata),
               data = STARdata,
               weights = STARdata_ate)
#> Warning: Some covariance adjustments are NA; be careful of dropping these observations when fitting the ITT effect model
coef(ate_read)
#>   (Intercept) treatmentTRUE 
#> -5.334478e-13  5.554179e+00 

ate_read_eth <- lm(readk ~ treatment * ethnicity,
                   offset = cov_adj(covariance_y0_read, newdata = STARdata),
                   data = STARdata,
                   weights = STARdata_ate)
#> Warning: Some covariance adjustments are NA; be careful of dropping these observations when fitting the ITT effect model
coef(ate_read_eth)
#>                     (Intercept)                   treatmentTRUE 
#>                   -4.599280e-13                    5.901494e+00 
#>                   ethnicityafam                  ethnicityasian 
#>                   -1.267935e-13                    5.281315e-13 
#>               ethnicityhispanic               ethnicityamindian 
#>                    6.750873e-12                    1.579858e-12 
#>                  ethnicityother     treatmentTRUE:ethnicityafam 
#>                    3.151440e-12                   -1.289929e+00 
#>    treatmentTRUE:ethnicityasian treatmentTRUE:ethnicityhispanic 
#>                   -2.494972e+01                    3.214437e+01 
#> treatmentTRUE:ethnicityamindian    treatmentTRUE:ethnicityother 
#>                    1.391799e+01                    3.202748e+00