Order observations used to fit a DirectAdjusted model and its covariance adjustment model

Order observations used to fit a DirectAdjusted model and its covariance adjustment model

Usage

.order_samples(x, by = NULL, ...)

Arguments

x

a fitted DirectAdjusted object

by

character vector; indicates unit of assignment columns to generate ID's from; default is NULL, which uses the unit of assignment columns specified in the DirectAdjusted object's Design slot

...

arguments passed to methods

Value

a list of four named vectors

Details

.order_samples underpins the ordering for .make_uoa_ids and estfun.DirectAdjusted. This function needs to order the rows in in \(\mathcal{Q}\cup\mathcal{C}\), but it also needs to explain how the original matrices of estimating equations should be indexed so the contributions from both sets of equations match. So instead of returning a numeric vector, which one might expect for an ordering function, .order_samples returns a list of vectors, which is explained in the Return section. Ultimately, the order is given by concatenating the vectors stored in Q_not_C, Q_in_C, and C_not_q (Q_in_C and C_in_Q are interchangeable in terms of deriving the order). The names of the Q_not_C and Q_in_C vectors correspond to row indices of the matrix of estimating equations for the ITT effect model, while the names of the C_in_Q and C_not_Q vectors correspond to row indices of the matrix of estimating equations for the covariance adjustment model. When a by argument is provided to cov_adj, it is used to deduce the order.